requirements.txt
requests
python-dotenv
tqdm
google-genai
fastapi
uvicorn
redis 
apscheduler